test(logic): split diplomatic suggestion API tests for file-size gate
Refs #2216. Split order suggestion API coverage into focused diplomatic test files so each file stays under the 400-line CI ceiling while preserving existing assertions.
